+++ docker/amster/export.sh
@@ -3,12 +3,10 @@
 
 set -x
 
-GIT_ROOT=${GIT_ROOT:-/git}
+GIT_ROOT=${GIT_ROOT:-/git/config}
 
 # This should be set by the downward API, but in case it isn't, default it.
 NAMESPACE=${NAMESPACE:-default}
-
-GIT_PROJECT_DIRECTORY="${GIT_PROJECT_DIRECTORY:-forgeops-init}"
 
 # Needed for any git ssh commands.
 export GIT_SSH_COMMAND="ssh -q -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no -i /etc/git-secret/ssh"
@@ -16,7 +14,7 @@ export GIT_SSH_COMMAND="ssh -q -o UserKnownHostsFile=/dev/null -o StrictHostKeyC
 # Default export path - relative to the root.
 export EXPORT_PATH="${EXPORT_PATH:-${NAMESPACE}/am/export}"
 
-cd "${GIT_ROOT}/${GIT_PROJECT_DIRECTORY}"
+cd "${GIT_ROOT}"
 
 git config core.filemode false
 git config user.email "auto-sync@forgerock.net"
@@ -25,7 +23,7 @@ git config user.name "Git Auto-sync user"
 git branch ${GIT_AUTOSAVE_BRANCH}
 git checkout ${GIT_AUTOSAVE_BRANCH}
 
-export AMSTER_EXPORT_PATH="${GIT_ROOT}/${GIT_PROJECT_DIRECTORY}/${EXPORT_PATH}"
+export AMSTER_EXPORT_PATH="${GIT_ROOT}/${EXPORT_PATH}"
 
 
 mkdir -p "${AMSTER_EXPORT_PATH}"