+++ docker/git/git-init.sh
@@ -2,11 +2,8 @@
 # Clone from git.
 set -x
 
-GIT_ROOT=${GIT_ROOT:=/git}
 
 GIT_BRANCH=${GIT_CHECKOUT_BRANCH:-master}
-GIT_PROJECT_DIRECTORY="${GIT_PROJECT_DIRECTORY:-forgeops-init}"
-
 
 
 export GIT_SSH_COMMAND="ssh -q -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no -i /etc/git-secret/ssh"
@@ -14,16 +11,18 @@ export GIT_SSH_COMMAND="ssh -q -o UserKnownHostsFile=/dev/null -o StrictHostKeyC
 # If GIT_REPO is defined, clone the configuration repo
 
 if [ ! -z "${GIT_REPO}" ]; then
-    mkdir -p ${GIT_ROOT}
+
+    mkdir -p "${GIT_ROOT}"
     cd ${GIT_ROOT}
     # sometimes the git repo emptyDir does not get cleaned up from a previous run
     rm -fr *
-    git clone "${GIT_REPO}"
+    git clone "${GIT_REPO}" "${GIT_ROOT}"
     if [ "$?" -ne 0 ]; then
-       echo "git clone failed"
+       echo "git clone failed. Will sleep for 5 minutes for debugging."
+       sleep 300
        exit 1
     fi
-    cd "${GIT_PROJECT_DIRECTORY}"
+    cd "${GIT_ROOT}"
     git checkout "${GIT_BRANCH}"
     if [ "$?" -ne 0 ]; then
        echo "git checkout of ${GIT_BRANCH} failed. Will sleep for 5 min for debugging"