+++ docker/openidm/git-sync.sh
@@ -1,10 +1,11 @@
 #!/usr/bin/env bash
 # Sync git configuration upstream. This assumes that the git project has already been cloned.
+# This is a convenience utility so we don't need to install another git sidecar.
+# TODO: Consider deprecating this.
 
 set -x
 # Top level directory where git projects will be cloned to.
-GIT_ROOT=${GIT_ROOT:=/git}
-GIT_PROJECT_DIRECTORY="${GIT_PROJECT_DIRECTORY:-forgeops-init}"
+GIT_ROOT=${GIT_ROOT:=/git/config}
 
 
 GIT_AUTOSAVE_BRANCH="${GIT_AUTOSAVE_BRANCH:-autosave-idm}"
@@ -13,7 +14,7 @@ GIT_AUTOSAVE_BRANCH="${GIT_AUTOSAVE_BRANCH:-autosave-idm}"
 export GIT_SSH_COMMAND="ssh -q -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no -i /etc/git-secret/ssh"
 
 
-cd "${GIT_ROOT}/${GIT_PROJECT_DIRECTORY}"
+cd "${GIT_ROOT}"
 
 # This configures git to ignore file mode changes.
 git config core.filemode false